Added SpkArchive file store for reading CoH:Online's data.spk archive (and hopefully other future SPK archives) Added IArchiveFileStore interface as a refinement of IFileStore and a base for SgaArchive and the new SpkArchive, moving some parts of SgaArchive down into it Improved automatic type deduction on CHECK_RANGE and CHECK_RANGE_LTMAX macros Added static doesFileResemble method to all archive classes to test if a file resembles that type of archive Added RainEmptyString as a global empty string constant The default constructor on a RainString now references the empty string rather than allocating a new empty buffer and RainString::clear also does this Added another overload of RainString::compareCaseless for comparison against ASCII string with known length Fixed file system store's date handling code Very minor optimisation to file system store's enumeration of present hard drives

+++ Rainman2/mem_fs.cpp
@@ -95,7 +95,7 @@ public:
 
   virtual void getItemDetails(size_t iIndex, directory_item_t& oDetails) throw(...)
   {
-    CHECK_RANGE_LTMAX(static_cast<size_t>(0), iIndex, getItemCount());
+    CHECK_RANGE_LTMAX(0, iIndex, getItemCount());
     if(iIndex < m_pDirectory->m_vSubdirectories.size())
     {
       MemoryFileStore::directory_t *pDirectory = m_pDirectory->m_vSubdirectories[iIndex];
@@ -369,7 +369,7 @@ size_t MemoryFileStore::getEntryPointCount() throw()
 
 const RainString& MemoryFileStore::getEntryPointName(size_t iIndex) throw(...)
 {
-  CHECK_RANGE_LTMAX(static_cast<size_t>(0), iIndex, m_vEntryPoints.size());
+  CHECK_RANGE_LTMAX(0, iIndex, m_vEntryPoints.size());
   return m_vEntryPoints[iIndex]->m_sName;
 }
